Core: Refactor the code into a scalable blockchain
