--- core/pow.go
@@ -1,1 +1,0 @@
-package core
