+++ internal/consensus/consensus.go
@@ -1,0 +1,1 @@
+package consensus
