+++ internal/consensus/pos.go
@@ -1,0 +1,1 @@
+package consensus
