+++ internal/consensus/pow.go
@@ -1,0 +1,1 @@
+package consensus
