+++ internal/core/block.go
