+++ internal/p2p/peer.go
@@ -1,0 +1,1 @@
+package p2p
