+++ tests/p2p_test.go
@@ -1,0 +1,1 @@
+package tests
